Trackign all allocs
It was hard to track package id at compile time for special allocs
(slices and maps for example).
As a result, I took a different approach that traces the call stack for
these cases.
We will see if this is a huge penalty or not for the performance.

I also had to avoid stack smashing for tracing (e.g., alloc inside the
tracer). For that I acquirem.