Adding a level of indirection to mcache
Now the mcache has a level of indirection that allows to have lists of
spans for a given class. Each element will be attributed to a single pkg
id.

Trying to use a map was not possible due to self cyclic dependency
(i.e., you need the allocator to allocate a map).
I might make it a sorted list.

Next: I need to instrument the central mcentrals to add the same level
of indirection.
The tiny allocator region needs to be handled too.