Introducing hooks for spans
I changed the runtime code to introduce hooks whenever a span gets
attributed a new id.

I also changed the backend.go file to be generic in terms of how it gets
initialized (in preparation for the multiple backend case).

@CharlyCst I put comments in mpk.go that you should read. We are
probably missing some global datastructures to efficiently implement the
register/transfer functions.
You should also check that the SView packages are also part of the SPkgs
slice inside the config.