IOC and KVM implementation + bug in mpk
The mpk assembly does not compile anymore and got commented out for the
moment (see with @CharlyCst).
I ported parts of IOC and KVM headers and constants and added a test to
check that we generate the correct values.
Maybe, later on, we should import the C values directly in the test and
do a comparison.