:warning: saving before refactoring
kvm requires a lot of setup if we want to be able to exploit long mode
w/ kernel and user spaces (which I am still not completely sure about to
be honest, but it seems to be the only way to have page tables).

I amd saving before refactoring the folder to see if I can reorganize
things.