Fixing rogue symbols in go runtime
The go linker generates a few symbols used for meta-access via the
runtime.
These symbols are crucial for the runtime to work, but are not marked as
belonging to the runtime package.
This patch assigns some of them to the runtime, bloats some of them, and
for those that cannot be assigned, it loads them from the elf (e.g.,
pclntab).