Fixing heap id attribution
There was an issue with dynamic allocation from a sandbox code block.
The id attributed to the span was the one of the surrounding package.
I fixed it by treating a sandbox as its own package.
This gave me the idea of implementing prisitine in the same way.
If the current goid is a sandbox one, then we should be able to state
that its id can be used to seggregate ALL its allocations, i.e.,
disregarding the package that actually did the allocation, and
attributing it to the current goid. This could work.