:warning: Dirty
Refactored the code so that we could have a single free space allocator
and PTEAllocator. This is easier to handler all the issues we had with
enclosures being able to update others.
I can now simplify the logic (i.e., remove oarenas) and then we can also
use the same god page table in all VMs.