Accepting nesting because I couldn't avoid it
There is a weird path that leads to a new package nested in another and
does not travel through our hooks. As a result, I merged the two
packages for the moment.
This appears with six and six.moves when doing
from skimage import data; i = data.camera().
I think it might have to do with the fucked up .so dependencies but not
sure.